remove Startup.cs
